restaurant-reservations/pyproject.toml
[project]
name = "restaurant-reservations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "dj-database-url>=2.2.0",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-filter>=24.2",
    "django-redis>=5.4.0",
    "djangorestframework>=3.16.1",
    "factory-boy>=3.3.0",
    "flower>=2.0.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "python-decouple>=3.8",
    "redis>=6.4.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["*Test"]
python_functions = ["test_*"]
addopts = [
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--reuse-db",
]
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "slow: Slow running tests",
]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]
